# resources/openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "Local server",
      "url": "http://localhost:8000"
    }
  ],
  "info": {
    "description": "TWID-SOA API for the resources service.",
    "version": "1.0.0",
    "title": "TWID-SOA-API-resources",
    "contact": {
      "email": "javiervidrua@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://jvr.mit-license.org/"
    }
  },
...
  "paths": {
...
    "/board/round": {
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "Get the round",
        "description": "Get the round",
        "responses": {
          "200": {
            "description": "Round number",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/board-round"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Board"
        ],
        "summary": "Create a new round (increment the current round number)",
        "description": "Create a new round (increment the current round number)",
        "responses": {
          "200": {
            "description": "Round created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/board-round"
                }
              }
            }
          },
          "400": {
            "description": "Cannot create any more rounds"
          }
        }
      },
      "delete": {
        "tags": [
          "Board"
        ],
        "summary": "Reset the round to round 1",
        "description": "Reset the round to round 1",
        "responses": {
          "200": {
            "description": "Round reset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/board-round"
                }
              }
            }
          }
        }
      }
    },
...
    "/board/score/{player}": {
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "Get the score of the specified player",
        "description": "Get the score of the specified player",
        "parameters": [
          {
            "in": "path",
            "name": "player",
            "description": "Name of the player",
            "required": true,
            "schema": {
              "type": "string",
              "example": "US"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Score of the specified player",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/board-score-player"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Board"
        ],
        "summary": "Update the score of the specified player",
        "description": "Update the score of the specified player",
        "parameters": [
          {
            "in": "path",
            "name": "player",
            "description": "Name of the player",
            "required": true,
            "schema": {
              "type": "string",
              "example": "US"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Score updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/board-score-player"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/board-score-player"
              }
            }
          },
          "description": "New score of the specified player"
        }
      },
      "delete": {
        "tags": [
          "Board"
        ],
        "summary": "Reset the score of a player to score 0",
        "description": "Reset the score of a player to score 0",
        "parameters": [
          {
            "in": "path",
            "name": "player",
            "description": "Name of the player",
            "required": true,
            "schema": {
              "type": "string",
              "example": "US"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Score reset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/board-score-player"
                }
              }
            }
          }
        }
      }
    },
...
    "/board/map/{region}": {
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "Get the countries of the specified region",
        "description": "Get the countries of the specified region",
        "parameters": [
          {
            "in": "path",
            "name": "region",
            "description": "Name of the region",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Africa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Countries of the specified region",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/board-map-region-country"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Board"
        ],
        "summary": "Update the countries of the specified region",
        "description": "Update the countries of the specified region",
        "parameters": [
          {
            "in": "path",
            "name": "region",
            "description": "Name of the region",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Africa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Countries of the specified region updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/board-map-region-country"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/board-map-region-country"
                }
              }
            }
          },
          "description": "New countries of the specified region"
        }
      }
    },
    "/board/map/{region}/{country}": {
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "Get the specified country of the specified region",
        "description": "Get the specified country of the specified region",
        "parameters": [
          {
            "in": "path",
            "name": "region",
            "description": "Name of the region",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Africa"
            }
          },
          {
            "in": "path",
            "name": "country",
            "description": "Name of the country",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Congo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Specified country of the specified region",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/board-map-region-country-country"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Board"
        ],
        "summary": "Update the specified country of the specified region",
        "description": "Update the specified country of the specified region",
        "parameters": [
          {
            "in": "path",
            "name": "region",
            "description": "Name of the region",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Africa"
            }
          },
          {
            "in": "path",
            "name": "country",
            "description": "Name of the country",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Congo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Specified country of the specified region updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/board-map-region-country-country"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/board-map-region-country-country"
              }
            }
          },
          "description": "New attributes of the specified country of the specified region"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "board-round": {
        "type": "object",
        "required": [
          "round"
        ],
        "properties": {
          "round": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 8,
            "example": 1
          }
        }
      },
...
      "board-score-player": {
        "type": "object",
        "required": [
          "score"
        ],
        "properties": {
          "score": {
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "maximum": 100,
            "example": 0
          }
        }
      },
...
      "board-map-region-country": {
        "type": "object",
        "required": [
          "country"
        ],
        "properties": {
          "country": {
            "type": "string",
            "example": "Congo"
          }
        }
      },
      "board-map-region-country-country": {
        "type": "object",
        "required": [
          "stability",
          "isConflictive",
          "isOilProducer",
          "influence"
        ],
        "properties": {
          "stability": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 5,
            "example": 2
          },
          "isConflictive": {
            "type": "boolean",
            "example": true
          },
          "isOilProducer": {
            "type": "boolean",
            "example": true
          },
          "influence": {
            "type": "object",
            "required": [
              "$player"
            ],
            "properties": {
              "$player": {
                "type": "integer",
                "format": "int32",
                "minimum": 0,
                "maximum": 100,
                "example": 1
              }
            }
          }
        },
        "example": {
          "stability": 2,
          "isConflictive": true,
          "isOilProducer": true,
          "influence": {
            "US": 1
          }
        }
      },
...
    }
  }
}